+++ example/lib/main.dart
@@ -1,16 +1,17 @@
+// ignore_for_file: avoid_print
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:html_editor_plus/html_editor.dart';
-import 'package:file_picker/file_picker.dart';
 
 import 'plus/example_scaffold.dart';
 
-void main() => runApp(HtmlEditorExampleApp(showPlusExample: false));
+void main() => runApp(const HtmlEditorExampleApp(showPlusExample: false));
 
 class HtmlEditorExampleApp extends StatelessWidget {
   final bool showPlusExample;
 
-  const HtmlEditorExampleApp({this.showPlusExample = false});
+  const HtmlEditorExampleApp({super.key, this.showPlusExample = false});
 
   @override
   Widget build(BuildContext context) => MaterialApp(
@@ -31,7 +32,7 @@ class HtmlEditorExample extends StatefulWidget {
   final String title;
 
   @override
-  _HtmlEditorExampleState createState() => _HtmlEditorExampleState();
+  State<HtmlEditorExample> createState() => _HtmlEditorExampleState();
 }
 
 class _HtmlEditorExampleState extends State<HtmlEditorExample> {
@@ -58,7 +59,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
           elevation: 0,
           actions: [
             IconButton(
-                icon: Icon(Icons.refresh),
+                icon: const Icon(Icons.refresh),
                 onPressed: () {
                   if (kIsWeb) {
                     _controller.reloadWeb();
@@ -72,7 +73,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
           onPressed: () {
             _controller.toggleCodeView();
           },
-          child: Text(r'<\>', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18)),
+          child: const Text(r'<\>', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18)),
         ),
         body: SingleChildScrollView(
           child: Column(
@@ -80,7 +81,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
             children: <Widget>[
               HtmlEditor(
                 controller: _controller,
-                htmlEditorOptions: HtmlEditorOptions(
+                htmlEditorOptions: const HtmlEditorOptions(
                   hint: 'Your text here...',
                   shouldEnsureVisible: true,
                   //initialText: "<p>text content initial, if any</p>",
@@ -108,7 +109,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                     return true;
                   },
                 ),
-                otherOptions: OtherOptions(height: 550),
+                otherOptions: const OtherOptions(height: 550),
                 callbacks: Callbacks(onBeforeCommand: (String? currentHtml) {
                   print('html before change is $currentHtml');
                 }, onChangeContent: (String? changed) {
@@ -188,9 +189,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.undo();
                       },
-                      child: Text('Undo', style: TextStyle(color: Colors.white)),
+                      child: const Text('Undo', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -198,9 +199,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.clear();
                       },
-                      child: Text('Reset', style: TextStyle(color: Colors.white)),
+                      child: const Text('Reset', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -208,7 +209,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                           backgroundColor: Theme.of(context).colorScheme.secondary),
                       onPressed: () async {
                         var txt = await _controller.getText();
-                        if (txt.contains('src=\"data:')) {
+                        if (txt.contains('src="data:')) {
                           txt =
                               '<text removed due to base-64 data, displaying the text could cause the app to crash>';
                         }
@@ -216,12 +217,12 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                           result = txt;
                         });
                       },
-                      child: Text(
+                      child: const Text(
                         'Submit',
                         style: TextStyle(color: Colors.white),
                       ),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -230,7 +231,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.redo();
                       },
-                      child: Text(
+                      child: const Text(
                         'Redo',
                         style: TextStyle(color: Colors.white),
                       ),
@@ -252,9 +253,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.disable();
                       },
-                      child: Text('Disable', style: TextStyle(color: Colors.white)),
+                      child: const Text('Disable', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -263,7 +264,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () async {
                         _controller.enable();
                       },
-                      child: Text(
+                      child: const Text(
                         'Enable',
                         style: TextStyle(color: Colors.white),
                       ),
@@ -271,7 +272,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                   ],
                 ),
               ),
-              SizedBox(height: 16),
+              const SizedBox(height: 16),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Row(
@@ -283,9 +284,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.insertText('Google');
                       },
-                      child: Text('Insert Text', style: TextStyle(color: Colors.white)),
+                      child: const Text('Insert Text', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -294,7 +295,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.insertHtml('''<p style="color: blue">Google in blue</p>''');
                       },
-                      child: Text('Insert HTML', style: TextStyle(color: Colors.white)),
+                      child: const Text('Insert HTML', style: TextStyle(color: Colors.white)),
                     ),
                   ],
                 ),
@@ -310,12 +311,12 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () async {
                         _controller.insertLink('Google linked', 'https://google.com', true);
                       },
-                      child: Text(
+                      child: const Text(
                         'Insert Link',
                         style: TextStyle(color: Colors.white),
                       ),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -326,7 +327,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                             'https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_92x30dp.png',
                             filename: 'Google network image');
                       },
-                      child: Text(
+                      child: const Text(
                         'Insert network image',
                         style: TextStyle(color: Colors.white),
                       ),
@@ -334,7 +335,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                   ],
                 ),
               ),
-              SizedBox(height: 16),
+              const SizedBox(height: 16),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Row(
@@ -345,9 +346,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.addNotification('Info notification', NotificationType.info);
                       },
-                      child: Text('Info', style: TextStyle(color: Colors.white)),
+                      child: const Text('Info', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -356,9 +357,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                         _controller.addNotification(
                             'Warning notification', NotificationType.warning);
                       },
-                      child: Text('Warning', style: TextStyle(color: Colors.white)),
+                      child: const Text('Warning', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -368,12 +369,12 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                         _controller.addNotification(
                             'Success notification', NotificationType.success);
                       },
-                      child: Text(
+                      child: const Text(
                         'Success',
                         style: TextStyle(color: Colors.white),
                       ),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -382,7 +383,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () {
                         _controller.addNotification('Danger notification', NotificationType.danger);
                       },
-                      child: Text(
+                      child: const Text(
                         'Danger',
                         style: TextStyle(color: Colors.white),
                       ),
@@ -390,7 +391,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                   ],
                 ),
               ),
-              SizedBox(height: 16),
+              const SizedBox(height: 16),
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Row(
@@ -402,9 +403,9 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                         _controller.addNotification(
                             'Plaintext notification', NotificationType.plaintext);
                       },
-                      child: Text('Plaintext', style: TextStyle(color: Colors.white)),
+                      child: const Text('Plaintext', style: TextStyle(color: Colors.white)),
                     ),
-                    SizedBox(
+                    const SizedBox(
                       width: 16,
                     ),
                     TextButton(
@@ -413,7 +414,7 @@ class _HtmlEditorExampleState extends State<HtmlEditorExample> {
                       onPressed: () async {
                         _controller.removeNotification();
                       },
-                      child: Text(
+                      child: const Text(
                         'Remove',
                         style: TextStyle(color: Colors.white),
                       ),